♻️ refactor: Allow Alert modal to close with Enter and Escape keys 🚚 rename: Move Alert component file to components/common directory
