+++ src/app/my-page/edit-profile/verify/VerifyForm.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Alert from '@/app/shop/[productType]/[id]/Alert';
+import Alert from '@/components/common/Alert';
 import Button from '@/components/common/Button';
 import Input from '@/components/common/Input';
 import { verifyEmail } from '@/data/actions/user';
